+++ sdk/common.py
@@ -1,0 +1,11 @@
+from private.memory.memory import Size
+
+# Constants
+BLOCK_SIZE = 1 * Size.MB
+ENCRYPTION_OVERHEAD = 28  # 16 bytes for AES-GCM tag, 12 bytes for nonce
+MIN_BUCKET_NAME_LENGTH = 3
+MIN_FILE_SIZE = 127  # 127 bytes
+
+# Exception classes
+class SDKError(Exception):
+    pass
